Match up movies
At present, there are only two movies -- Kooky, and Indie Game: The
Movie -- but in my annex they get handled a little differently from
games. To wit, the Humble Bundle people put the link in each OS, which
is silly, because movies aren't games -- they're more like books, a
separate kind of object. Additionally, they're offered as .zip files,
which should be unpacked to the juicy media file within. However, we
can't know what they'll unpack to until we download them, so create
yet another mapping dict to provide the name of the contained
file. Movies seem to never be updated, so we should never have to
redownload an existing movie file, so if the contained file exists on
the local filesystem, we should be in the clear.

More of this code should be shared between movies and books.