More plausible resolve_missing logic
There's a lot of weird corner cases, especially in the places where
there's only one file locally. Humble Bundle adds new downloads and
renames the old ones, so you get to try to figure out what everything
is. In this (important) case, try to match a file which doesn't
correspond clearly to any of the downloads to all the downloads
available. This requires that the loop be much more complicated.

Pull out find_closest_file as a separate method. Change structure of
the loop to be a while. If either file_list or hdl_list is empty,
we're clear.

Change does_match to not return UserInvestigate; let resolve_missing
generate UserInvestigate actions if it can't find anything that looks
right.