Support comedy
Comedy videos are offered as .zip files with (usually) a single video
file inside. Download and unpack them. To recognize them as the same
file later, invent a new FileMatch class that says "This is the same
file so don't redownload it."

This code probably fails in some edge cases but it seems to work OK now.